+++ browser/themes/shared/autocomplete.inc.css
@@ -164,7 +164,7 @@
 }
 
 #PopupAutoComplete > richlistbox > richlistitem[originaltype="insecureWarning"] > .ac-site-icon {
-  list-style-image: url(chrome://global/skin/icons/connection-mixed-active-loaded.svg);
+  list-style-image: url(chrome://browser/skin/connection-mixed-active-loaded.svg);
   -moz-context-properties: fill;
   fill: var(--grey-60);
 }
