+++ browser/themes/shared/identity-block/identity-block.inc.css
@@ -167,14 +167,14 @@
 #identity-box[pageproxystate="valid"].mixedDisplayContentLoadedActiveBlocked > #identity-icon,
 #identity-box[pageproxystate="valid"].certUserOverridden > #identity-icon,
 #identity-box[pageproxystate="valid"].certErrorPage > #identity-icon {
-  list-style-image: url(chrome://global/skin/icons/connection-mixed-passive-loaded.svg);
+  list-style-image: url(chrome://browser/skin/connection-mixed-passive-loaded.svg);
   fill: unset;
 }
 
 #identity-box[pageproxystate="valid"].notSecure > #identity-icon,
 #identity-box[pageproxystate="valid"].mixedActiveContent > #identity-icon,
 #identity-box[pageproxystate="valid"].httpsOnlyErrorPage > #identity-icon {
-  list-style-image: url(chrome://global/skin/icons/connection-mixed-active-loaded.svg);
+  list-style-image: url(chrome://browser/skin/connection-mixed-active-loaded.svg);
 }
 
 #permissions-granted-icon {
