+++ toolkit/components/telemetry/moz.build
@@ -80,8 +80,6 @@ SOURCES += [
 if CONFIG["OS_ARCH"] == "WINNT":
     SOURCES += [
         "other/UntrustedModules.cpp",
-        "other/UntrustedModulesBackupService.cpp",
-        "other/UntrustedModulesDataSerializer.cpp",
     ]
 
 XPCOM_MANIFESTS += [
