+++ toolkit/content/aboutSupport.xhtml
@@ -784,25 +784,6 @@
       </table>
 #endif
 
-#ifdef XP_WIN
-      <h2 id="third-party-modules" class="major-section" hidden="true"
-          data-l10n-id="support-third-party-modules-title"/>
-      <table id="third-party-modules-table" class="prefs-table" hidden="true">
-        <thead id="third-party-modules-thead">
-          <th data-l10n-id="support-third-party-modules-module"/>
-          <th class="third-party-modules-column-version"
-              data-l10n-id="support-third-party-modules-version"/>
-          <th data-l10n-id="support-third-party-modules-vendor"/>
-          <th class="third-party-modules-header-occurrence" colspan="2"
-              data-l10n-id="support-third-party-modules-occurrence"/>
-        </thead>
-        <tbody id="third-party-modules-tbody">
-          <tr><td id="third-party-modules-no-data" colspan="5" hidden="true"
-                  data-l10n-id="support-third-party-modules-empty"></td></tr>
-        </tbody>
-      </table>
-#endif
-
     </div>
 
   </body>
