+++ toolkit/modules/tests/browser/browser_Troubleshoot.js
@@ -980,12 +980,6 @@ const SNAPSHOT_SCHEMA = {
         },
       },
     },
-    thirdPartyModules: {
-      required:
-        AppConstants.platform == "win" &&
-        Services.prefs.getBoolPref("browser.enableAboutThirdParty"),
-      type: "array",
-    },
   },
 };
