+++ toolkit/xre/UntrustedModulesData.h
@@ -189,7 +189,6 @@ class UntrustedModulesData final {
   void AddNewLoads(const ModulesMap& aModulesMap,
                    Vector<ProcessedModuleLoadEvent>&& aEvents,
                    Vector<Telemetry::ProcessedStack>&& aStacks);
-  void Merge(UntrustedModulesData&& aNewData);
 
   void Swap(UntrustedModulesData& aOther);
