--- toolkit/xre/test/gtest/TestUntrustedModules_Dll2/moz.build
@@ -1,15 +1,0 @@
-# -*- Mode: python; indent-tabs-mode: nil; tab-width: 40 -*-
-# This Source Code Form is subject to the terms of the Mozilla Public
-# License, v. 2.0. If a copy of the MPL was not distributed with this
-# file, You can obtain one at http://mozilla.org/MPL/2.0/.
-
-DIST_INSTALL = False
-
-SharedLibrary("TestUntrustedModules_Dll2")
-
-UNIFIED_SOURCES = [
-    "TestUntrustedModules_Dll2.cpp",
-]
-
-if CONFIG["COMPILE_ENVIRONMENT"]:
-    TEST_HARNESS_FILES.gtest += ["!TestUntrustedModules_Dll2.dll"]
